+++ app/containers/App/index.js
@@ -11,7 +11,7 @@ import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 
 import HomePage from 'containers/HomePage/Loadable';
-import ManagerPage from 'containers/ManagerPage/Loadable'
+import MngrPage from 'containers/MngrPage/Loadable'
 import NotFoundPage from 'containers/NotFoundPage/Loadable';
 
 import GlobalStyle from '../../global-styles';
@@ -21,7 +21,7 @@ export default function App() {
     <div>
       <Switch>
         <Route exact path="/" component={HomePage} />
-        <Route exact path="/management" component={ManagerPage} />
+        <Route exact path="/management" component={MngrPage} />
         <Route component={NotFoundPage} />
       </Switch>
       <GlobalStyle />
