--- app/containers/ManagerPage/actions.js
@@ -1,9 +1,0 @@
-import { LOAD_RADIO_DATA } from './constants'
-
-
-export function loadRadioData(btnPressed){
-    return {
-        type: LOAD_RADIO_DATA,
-        btnPressed
-    }
-}
