--- app/containers/ManagerPage/constants.js
@@ -1,1 +1,0 @@
-export const LOAD_RADIO_DATA = 'app/ManagerPage/LOAD_RADIO_DATA'
