+++ app/containers/MngrPage/Loadable.js
@@ -1,6 +1,9 @@
 /**
- * Asynchronously loads the component for HomePage
+ *
+ * Asynchronously loads the component for MngrPage
+ *
  */
+
 import loadable from 'loadable-components';
 
 export default loadable(() => import('./index'));
