+++ app/containers/MngrPage/constants.js
@@ -1,0 +1,9 @@
+/*
+ *
+ * MngrPage constants
+ *
+ */
+//Actions to get radio button needed data
+export const GET_RADIO_DATA = 'app/MngrPage/GET_RADIO_DATA';
+export const GET_RADIO_DATA_SUCCESS = 'app/MngrPage/GET_RADIO_DATA';
+export const GET_RADIO_DATA_ERROR = 'app/MngrPage/GET_RADIO_DATA';
