+++ app/containers/MngrPage/messages.js
@@ -1,0 +1,16 @@
+/*
+ * MngrPage Messages
+ *
+ * This contains all the text for the MngrPage container.
+ */
+
+import { defineMessages } from 'react-intl';
+
+export const scope = 'app.containers.MngrPage';
+
+export default defineMessages({
+  header: {
+    id: `${scope}.header`,
+    defaultMessage: 'This is the MngrPage container!',
+  },
+});
