+++ app/containers/MngrPage/tests/index.test.js
@@ -1,0 +1,11 @@
+// import React from 'react';
+// import { mount } from 'enzyme';
+// import { enzymeFind } from 'styled-components/test-utils';
+
+// import { MngrPage } from '../index';
+
+describe('<MngrPage />', () => {
+  it('Expect to have unit tests specified', () => {
+    expect(true).toEqual(false);
+  });
+});
