+++ app/containers/MngrPage/tests/reducer.test.js
@@ -1,0 +1,8 @@
+import { fromJS } from 'immutable';
+import mngrPageReducer from '../reducer';
+
+describe('mngrPageReducer', () => {
+  it('returns the initial state', () => {
+    expect(mngrPageReducer(undefined, {})).toEqual(fromJS({}));
+  });
+});
