+++ app/containers/MngrPage/tests/selectors.test.js
@@ -1,0 +1,8 @@
+// import { fromJS } from 'immutable';
+// import { selectMngrPageDomain } from '../selectors';
+
+describe('selectMngrPageDomain', () => {
+  it('Expect to have unit tests specified', () => {
+    expect(true).toEqual(false);
+  });
+});
